// src/core/domain/transactions/enums/availableAnalyticMetric.enum.ts
export enum AvailableAnalyticMetric {
  TRANSACTIONS_COUNT_BY_CATEGORY_AND_DATE_RANGE = 0,
  TRANSACTIONS_SUMT_BY_CATEGORY_AND_DATE_RANGE = 1,
  TRANSACTIONS_COUNT_BY_DATE_RANGE = 2,
  TRANSACTIONS_SUMT_BY_DATE_RANGE = 3,
  TRANSACTIONS_COUNT_RATIO_BY_CATEGORY_AND_DATE_RANGE = 4,
  TRANSACTIONS_SUM_RATIO_BY_CATEGORY_AND_DATE_RANGE = 5,
  TRANSACTIONS_COUNT_CHANGE_BY_CATEGORY_AND_DATE_RANGE = 6,
  TRANSACTIONS_SUM_CHANGE_BY_CATEGORY_AND_DATE_RANGE = 7,
}